make maven publishing work
After an unfruitful attempt at trying to publish to Maven Central
with the gradle-maven-plugin (which does not seem to support it), this
change fully uses and configures the gradle-maven-publish-plugin [1].

However, it seems, that publishing SNAPSHOT dependencies using this
plugin is not possible as of yet. It seems, if we'd be using maven,
we could use the maven-publishing-plugin. However in that case,
dokka would not be as simple to setup and work as flawlessly
as it does seem to do now.

Therefore, we bite the bullet and if the need arises to publish
SNAPSHOT builds, we'll use a different repository or waste more
time wading through old docs and outdated internet links, to make
it work somehow.

[1]: https://vanniktech.github.io/gradle-maven-publish-plugin/central/
[2]: https://central.sonatype.org/publish/publish-portal-maven/